+++ test/unit/equality_test.exs
@@ -72,17 +72,27 @@ defmodule RDF.EqualityTest do
     ]
     @value_unequal_strings [
     ]
+    @value_equal_strings_by_coercion [
+      {RDF.string("foo"), "foo"},
+    ]
+    @value_unequal_strings_by_coercion [
+      {RDF.string("foo"), "bar"},
+    ]
     @incomparable_strings [
+      {RDF.string("42"), 42},
+      {RDF.lang_string("foo", language: "de"), "foo"},
       {RDF.string("foo"), RDF.lang_string("foo", language: "de")},
       {RDF.lang_string("foo", language: "de"), RDF.string("foo")},
       {RDF.string("foo"), RDF.bnode("foo")},
     ]
 
-    test "term equality",    do: assert_term_equal    @term_equal_strings
-    test "term inequality",  do: assert_term_unequal  @term_unequal_strings
-    test "value equality",   do: assert_value_equal   @value_equal_strings
-    test "value inequality", do: assert_value_unequal @value_unequal_strings
-    test "incomparability",  do: assert_incomparable  @incomparable_strings
+    test "term equality",            do: assert_term_equal    @term_equal_strings
+    test "term inequality",          do: assert_term_unequal  @term_unequal_strings
+    test "value equality",           do: assert_value_equal   @value_equal_strings
+    test "value inequality",         do: assert_value_unequal @value_unequal_strings
+    test "coerced value equality",   do: assert_value_equal   @value_equal_strings_by_coercion
+    test "coerced value inequality", do: assert_value_unequal @value_unequal_strings_by_coercion
+    test "incomparability",          do: assert_incomparable  @incomparable_strings
   end
 
   describe "RDF.Boolean" do
@@ -110,18 +120,28 @@ defmodule RDF.EqualityTest do
       # invalid literals
       {RDF.boolean("foo"), RDF.boolean("bar")},
     ]
+    @value_equal_booleans_by_coercion [
+      {RDF.true,  true},
+      {RDF.false, false},
+    ]
+    @value_unequal_booleans_by_coercion [
+      {RDF.true,  false},
+      {RDF.false, true},
+    ]
     @incomparable_booleans [
-      {RDF.true,       nil},
-      {nil,            RDF.true},
-      {RDF.true,       RDF.string("FALSE")},
-      {RDF.integer(0), RDF.true},
+      {RDF.false, nil},
+      {RDF.true,  42},
+      {RDF.true,  RDF.string("FALSE")},
+      {RDF.true,  RDF.integer(0)},
     ]
 
-    test "term equality",    do: assert_term_equal    @term_equal_booleans
-    test "term inequality",  do: assert_term_unequal  @term_unequal_booleans
-    test "value equality",   do: assert_value_equal   @value_equal_booleans
-    test "value inequality", do: assert_value_unequal @value_unequal_booleans
-    test "incomparability",  do: assert_incomparable  @incomparable_booleans
+    test "term equality",            do: assert_term_equal    @term_equal_booleans
+    test "term inequality",          do: assert_term_unequal  @term_unequal_booleans
+    test "value equality",           do: assert_value_equal   @value_equal_booleans
+    test "value inequality",         do: assert_value_unequal @value_unequal_booleans
+    test "coerced value equality",   do: assert_value_equal   @value_equal_booleans_by_coercion
+    test "coerced value inequality", do: assert_value_unequal @value_unequal_booleans_by_coercion
+    test "incomparability",          do: assert_incomparable  @incomparable_booleans
   end
 
   describe "RDF.Numeric" do
@@ -162,16 +182,33 @@ defmodule RDF.EqualityTest do
       {RDF.decimal("foo"), RDF.decimal("bar")},
       {RDF.double("foo"),  RDF.double("bar")},
     ]
+    @value_equal_numerics_by_coercion [
+      {RDF.integer(42),   42},
+      {RDF.integer(42),   42.0},
+      {RDF.decimal(42),   42},
+      {RDF.decimal(3.14), 3.14},
+      {RDF.double(42),    42},
+      {RDF.double(3.14),  3.14},
+    ]
+    @value_unequal_numerics_by_coercion [
+      {RDF.integer(3),    3.14},
+      {RDF.double(3.14),  3},
+      {RDF.decimal(3.14), 3},
+    ]
     @incomparable_numerics [
-      {RDF.string("42"),  RDF.integer(42)},
+      {RDF.integer("42"), nil},
+      {RDF.integer("42"), true},
+      {RDF.integer("42"), "42"},
       {RDF.integer("42"), RDF.string("42")},
     ]
 
-    test "term equality",    do: assert_term_equal    @term_equal_numerics
-    test "term inequality",  do: assert_term_unequal  @term_unequal_numerics
-    test "value equality",   do: assert_value_equal   @value_equal_numerics
-    test "value inequality", do: assert_value_unequal @value_unequal_numerics
-    test "incomparability",  do: assert_incomparable  @incomparable_numerics
+    test "term equality",            do: assert_term_equal    @term_equal_numerics
+    test "term inequality",          do: assert_term_unequal  @term_unequal_numerics
+    test "value equality",           do: assert_value_equal   @value_equal_numerics
+    test "value inequality",         do: assert_value_unequal @value_unequal_numerics
+    test "coerced value equality",   do: assert_value_equal   @value_equal_numerics_by_coercion
+    test "coerced value inequality", do: assert_value_unequal @value_unequal_numerics_by_coercion
+    test "incomparability",          do: assert_incomparable  @incomparable_numerics
   end
 
   describe "RDF.DateTime" do
@@ -182,7 +219,7 @@ defmodule RDF.EqualityTest do
       {RDF.date_time("foo"), RDF.date_time("foo")},
     ]
     @term_unequal_datetimes [
-     {RDF.date_time("2002-04-02T12:00:00"), RDF.date_time("2002-04-02T17:00:00")},
+      {RDF.date_time("2002-04-02T12:00:00"), RDF.date_time("2002-04-02T17:00:00")},
       # invalid literals
       {RDF.date_time("foo"), RDF.date_time("bar")},
     ]
@@ -200,18 +237,96 @@ defmodule RDF.EqualityTest do
       # invalid literals
       {RDF.date_time("foo"), RDF.date_time("bar")},
     ]
+    @value_equal_datetimes_by_coercion [
+      {RDF.date_time("2002-04-02T12:00:00-01:00"), elem(DateTime.from_iso8601("2002-04-02T12:00:00-01:00"), 1)},
+      {RDF.date_time("2002-04-02T12:00:00"), ~N"2002-04-02T12:00:00"},
+    ]
+    @value_unequal_datetimes_by_coercion [
+      {RDF.date_time("2002-04-02T12:00:00-01:00"), elem(DateTime.from_iso8601("2002-04-02T12:00:00+00:00"), 1)},
+    ]
     @incomparable_datetimes [
       {RDF.string("2002-04-02T12:00:00-01:00"),    RDF.date_time("2002-04-02T12:00:00-01:00")},
-      {RDF.date_time("2002-04-02T12:00:00-01:00"), RDF.string("2002-04-02T12:00:00-01:00")},
     ]
 
-    test "term equality",    do: assert_term_equal    @term_equal_datetimes
-    test "term inequality",  do: assert_term_unequal  @term_unequal_datetimes
-    test "value equality",   do: assert_value_equal   @value_equal_datetimes
-    test "value inequality", do: assert_value_unequal @value_unequal_datetimes
-    test "incomparability",  do: assert_incomparable  @incomparable_datetimes
+    test "term equality",            do: assert_term_equal    @term_equal_datetimes
+    test "term inequality",          do: assert_term_unequal  @term_unequal_datetimes
+    test "value equality",           do: assert_value_equal   @value_equal_datetimes
+    test "value inequality",         do: assert_value_unequal @value_unequal_datetimes
+    test "coerced value equality",   do: assert_value_equal   @value_equal_datetimes_by_coercion
+    test "coerced value inequality", do: assert_value_unequal @value_unequal_datetimes_by_coercion
+    test "incomparability",          do: assert_incomparable  @incomparable_datetimes
   end
 
+  describe "RDF.Date" do
+    @term_equal_dates [
+      {RDF.date("2002-04-02-01:00"), RDF.date("2002-04-02-01:00")},
+      {RDF.date("2002-04-02"),       RDF.date("2002-04-02")},
+      # invalid literals
+      {RDF.date("foo"), RDF.date("foo")},
+    ]
+    @term_unequal_dates [
+      {RDF.date("2002-04-01"), RDF.date("2002-04-02")},
+      # invalid literals
+      {RDF.date("foo"), RDF.date("bar")},
+    ]
+    @value_equal_dates [
+    ]
+    @value_unequal_dates [
+    ]
+    @value_equal_dates_by_coercion [
+      {RDF.date("2002-04-02"), Date.from_iso8601!("2002-04-02")},
+    ]
+    @value_unequal_dates_by_coercion [
+      {RDF.date("2002-04-02"), Date.from_iso8601!("2002-04-03")},
+    ]
+    @incomparable_dates [
+      {RDF.date("2002-04-02"), RDF.string("2002-04-02")},
+    ]
+
+    test "term equality",            do: assert_term_equal    @term_equal_dates
+    test "term inequality",          do: assert_term_unequal  @term_unequal_dates
+    test "value equality",           do: assert_value_equal   @value_equal_dates
+    test "value inequality",         do: assert_value_unequal @value_unequal_dates
+    test "coerced value equality",   do: assert_value_equal   @value_equal_dates_by_coercion
+    test "coerced value inequality", do: assert_value_unequal @value_unequal_dates_by_coercion
+    test "incomparability",          do: assert_incomparable  @incomparable_dates
+  end
+
+  describe "RDF.Time" do
+    @term_equal_times [
+      {RDF.time("12:00:00+01:00"), RDF.time("12:00:00+01:00")},
+      {RDF.time("12:00:00"),       RDF.time("12:00:00")},
+      # invalid literals
+      {RDF.time("foo"), RDF.time("foo")},
+    ]
+    @term_unequal_times [
+      {RDF.time("12:00:00"), RDF.time("13:00:00")},
+      # invalid literals
+      {RDF.time("foo"), RDF.time("bar")},
+    ]
+    @value_equal_times [
+    ]
+    @value_unequal_times [
+    ]
+    @value_equal_times_by_coercion [
+      {RDF.time("12:00:00"), Time.from_iso8601!("12:00:00")},
+    ]
+    @value_unequal_times_by_coercion [
+      {RDF.time("12:00:00"), Time.from_iso8601!("13:00:00")},
+    ]
+    @incomparable_times [
+      {RDF.time("12:00:00"), RDF.string("12:00:00")},
+    ]
+
+    test "term equality",            do: assert_term_equal    @term_equal_times
+    test "term inequality",          do: assert_term_unequal  @term_unequal_times
+    test "value equality",           do: assert_value_equal   @value_equal_times
+    test "value inequality",         do: assert_value_unequal @value_unequal_times
+    test "coerced value equality",   do: assert_value_equal   @value_equal_times_by_coercion
+    test "coerced value inequality", do: assert_value_unequal @value_unequal_times_by_coercion
+    test "incomparability",          do: assert_incomparable  @incomparable_times
+  end
+  
   describe "RDF.Literals with unsupported types" do
     @equal_literals [
       {RDF.literal("foo", datatype: "http://example.com/datatype"),
@@ -264,6 +379,15 @@ defmodule RDF.EqualityTest do
       to be:   #{inspect expected}
       but got: #{inspect result}
       """
+
+    result = RDF.Term.equal?(right, left)
+    assert result == expected, """
+    expected RDF.Term.equal?(
+      #{inspect right},
+      #{inspect left})
+    to be:   #{inspect expected}
+    but got: #{inspect result}
+    """
   end
 
   defp assert_value_equality({left, right}, expected) do
@@ -275,6 +399,15 @@ defmodule RDF.EqualityTest do
       to be:   #{inspect expected}
       but got: #{inspect result}
       """
+
+    result = RDF.Term.equal_value?(right, left)
+    assert result == expected, """
+    expected RDF.Term.equal_value?(
+      #{inspect right},
+      #{inspect left})
+    to be:   #{inspect expected}
+    but got: #{inspect result}
+    """
   end
 
 end