+++ CHANGELOG.md
@@ -11,7 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 
 - top-level alias functions for constructors of the basic datatypes
 - top-level constant functions `RDF.true` and `RDF.false` for the two boolean 
-  RDF.Literal values
+  `RDF.Literal` values
 - `RDF.Decimal` datatype for `xsd:decimal` literals and support for decimal 
   literals in Turtle encoder
 - `RDF.Numeric` with a list of all numeric datatypes and shared functions for
@@ -22,6 +22,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
   from the XPath and SPARQL specs on `RDF.Boolean`
 - various functions on the `RDF.DateTime` and `RDF.Time` datatypes
 - `RDF.term?/1` 
+- `RDF.Term.coerce/1` which converts native Elixir values to a proper RDF term
 - `RDF.Term.equal?/2` and `RDF.Term.equal_value?/2`
 - `RDF.LangString.match_language?/2`
 - possibility to configure an application-specific default base IRI; for now it